+++ packages/orq-rc/src/models/operations/createpromptsnippet.ts
@@ -6200,7 +6200,7 @@ export const CreatePromptSnippetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-02-05T19:04:45.364Z",
+    "2025-02-05T22:03:39.496Z",
   ).transform(v => new Date(v)),
   type: CreatePromptSnippetPromptSnippetsType$inboundSchema,
   versions: z.array(z.lazy(() => Versions$inboundSchema)),
@@ -6249,7 +6249,7 @@ export const CreatePromptSnippetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string(),
   updatedById: z.string(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-02-05T19:04:45.364Z"))
+  updated: z.date().default(() => new Date("2025-02-05T22:03:39.496Z"))
     .transform(v => v.toISOString()),
   type: CreatePromptSnippetPromptSnippetsType$outboundSchema,
   versions: z.array(z.lazy(() => Versions$outboundSchema)),