+++ packages/orq-rc/RELEASES.md
@@ -189,3 +189,13 @@ Based on:
 - [typescript v3.1.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.39 - packages/orq-rc
+
+## 2025-02-05 22:04:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.484.1 (2.503.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.40 - packages/orq-rc